Added tutorial document.
This tutorial was originally created as a gift for a 6th-grade student on a flash drive, meant to enable basic python programming without installing to the computer's hard drive. Formatting the tutorial in word made the most sense at the time; it may be better to format the tutorial as markdown so it can be read in the browser directly.